+++ src/js/pages/index.js
@@ -29,6 +29,7 @@ export { default as PlaylistDetail } from './PlaylistDetail';
 
 export { default as Settings } from './Settings';
 export { default as SettingsAppearance } from './SettingsAppearance';
+export { default as SettingsChangelog } from './SettingsChangelog';
 export { default as SettingsOptions } from './SettingsOptions';
 
 export { default as Error404Auth } from './Error404Auth';